Precompute zero shifting of interger representations
Instead of subtracting the zero representation from each digit, we
now subtract the appropriate amount from the sum. That is, for a 3-digit
number, we use `111 * zero` as the shifting factor. We use a switch
statement to choose between the three options: 3, 4, or 5 digit.

We also did a quick analysis that showed that most of the numbers are
4 digit numbers.