Switch compilers, try out multithreading
Tried using multithreading using OpenMP, which doesn't require much
changing of the code. In this process, I had to switch compilers. The
good new is, this sped it up. The bad news is that it was the new
compiler, not the multithreading that sped it up. I guess there was
just too much overhead.